Established Art File Structure + Prototype Plinth Model
All art assets for the prototype will be included in "Art -> TempA_PrototypeArt1" (if Daniel and Kalina want the name changed, that's fine).
Once the prototype is complete, we'll create a new folder for final art assets.

This commit also includes the Plinth model/base texture from my Art Test (for the prototype, this is a generic square piece, as the rounded Plinth models in the Art Test would require a new system that generates the level's Plinth meshes using modular pieces, a purely aesthetic feature that isn't required for the prototype. IF YOU WANT THE MODULAR PLINTH MESHES, please let me (Ben Gregg - 1903707) know, and I'll commit them to this branch ).

Also, the Plinth model is supposed to be slightly bigger than 1 tile (it just looked nice in the Art Test, but a more suitable grid-sized piece can be designed for the final product).