Commit: - Deleted the old Art Test Assets (except the Display Stand assets) and edited the old Contemporary Art Assets by making each asset singular.
I have deleted all but the Display Stand Assets from the Art Test that will have no input in the final build. I also deleted the textures for the Art Test assets except for the Display Stands as well as it was causing confusion and taking up space. I inputted the following assets individually as OBJ's instead: - CurvedWall 01, 02 & 03, Entrance 01 & 02, Pillar 01, 02 & 03, SlantedWall 01, ThickCeiling 01, ThinCeiling 01, ThickWall 01, 02 & 03 and ThinWall 01 & 02. The textures for these new assets are still in the project and can still be applied.