Commit: - 1 Plinth Model with Glass Case Asset and 1 Display Stand Asset
I have made two new models to use in the Museum space to display Museum Artefacts. The asset “TempA_Plinth_01” is 1x1x1 metres and “TempA_Glass_P_01” is 2x1x1 metres. The asset “TempA_DisplayStand_01” is 3x3x3 metres. There is a Glass material in the folder “DisplayStands” that can be used for each of the glass aspects on both models. The “TempA_Glass_P_01” asset is removable, so there is the option to just having the “TempA_Plinth_01” asset standalone.