Commit: - To Temp_Art_2 Branch - Create a few Folders and Imported Dinosaur Footprint Model
I have created a few folders in the Artefacts folder. This is the pathway: - “Assets -> Art -> Artefacts”. The new folders are “Ceiling” and “Pillar”. I also imported in the K1972_285_Dinosaur_Footprint Asset Model with textures and a material. This asset is in the appropriate folders setup in the Temp_Art_2 Branch. This is the pathway to the asset and textures: - Assets -> Art -> Environment -> Static -> Models AND Materials. The folders have been named with the ID number and asset name.