Commit: - To Main Branch: - Imported 8 Optimised 3D Artefact Assets and their associating textures and removed older version of Inchyra Stone Asset.
The models and their associating textures have been allocated in proper folder locations adhering to the Art Standardisation document. The pathway is this: - Assets -> Art -> Artifacts -> (Then there are four folders in where the Assets and Textures have been distributed accordingly.

Here as follows are the models imported into the Project.: -
1.	DMA_5-1945_Inchyra-Stone – Located in the Misc Folder
2.	DMA_1979-5_Pictish-Cross-Slab-Fragment – Located in the Misc Folder
3.	DMA_1992-503_Celtic-Stone-Head – Located in the Misc Folder
4.	DMA_1992-815_St-Madoes-Stone – Located in the Misc Folder
5.	DMA_2000-43_Zoomorphic-Finial – Located in the Misc Folder
6.	DMA_2005-58_8th-9th-Century-Penannular-Brooch-Fragment – Located in the Misc Folder
7.	DMA_2151_Perth-Mirror-Case – Located in the Misc Folder
8.	DMA_K1972-285_Dinosaur-Footprint – Located in the Floor Folder

There is another asset to be imported as well, but I need to fix the name of it again as it is incorrect.