Commit: - In Main Branch – Descriptions of Artefacts
The descriptions included are of all the 3D Scanned Assets provided from the Client, the Assets produced from using the Image Collection and the Painting Images from the Image Collection as well.

Pathway: - “Assets” -> “ArtefactData” -> “ArtefactDescriptions”

There are two folders, separating the descriptions for Artefact Assets in the Prototype Build and those that are not. I have inputted in the Artefact Name the same name of the asset with “_Description”.