added improved cam and frame packages from DC
asset scriptable objects now have a variable for pixel sizes if its a painting, default to 0. looked at the inspector script and figured out how things are activated. seems simple enough but with it being tied to the camera controller, will have to update that first before continuing with model viewer