Updated Editor Camera and Added URP compatible Blur
Editor camera now has a threshold when checking if the camera has reached it's target when switching to perspective mode.
Also it now returns the user so they are the same distance away as they were in Edit mode and at the same pitch, but uses the Y rotation from where you ended up looking and your position from perspective mode.
URP blur taken from https://github.com/tomc128/urp-kawase-blur and modified to suit our needs/optimise. When fading in the ModelViewer changes the alpha value of the effect rather than the blur steps (of which are fixed in the Forward Renderer Feature that has been added to Art/UniversalRenderPipelineAsset_Renderer)