creating frame UI
so for the edit ui to handle paintings and frame swapping, i had to create new folders for the game to access the frame and paintings prefabs at run time. this needed to be done anyway for the other ui lists. i have started assembling the ui list for displaying the different frames. the next step wll be to get the ui to actually show the items and then after that to have the frame and paintings be swapped on button press. i also changed the project settings to only allow landscape