able to turn off and on ui
before certain ui's wouldn't turn off when entering edit and perspective mode. also menus would stay on when on the Main menu screen. thats all been fixed now being handled by the camera. tried also to include snap zones in the asset saving so that when we reload the asset it will set the snap zone its attached to to be invalid but with the current state of saving to networks i couldn't test it